Fix ordering in batch requests greater than 10.

+++ rest/rest-sqs/src/main/scala/org/elasticmq/rest/sqs/BatchRequestsModule.scala
@@ -73,6 +73,6 @@ object BatchRequestsModule {
         }
     }
 
-    subParameters.toList.sortBy(_._1).map(_._2)
+    subParameters.toList.sortBy(_._1.toInt).map(_._2)
   }
 }
